requirements.txt
asgiref==3.4.1
Django==3.2.8
django-crispy-forms==1.13.0
mysqlclient==2.0.3
phonenumbers==8.12.35
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
